Add style to greencurry and change style index
The style of the index page interfered with the style of the green curry
page, and had to be corrected. Added a home link in the greencurry file
that navigates towards the index file, I styled the file as well.

Used a lot of different selectors to play around with the cascade and
to show that I understand when to use what type of selector.